+++ Starbucks/Starbucks/Model/EventDataManager.swift
@@ -6,11 +6,18 @@
 //
 
 import Foundation
+import Combine
 
 class EventDataManager {
 
+    static let shared = EventDataManager()
+    let notifyEvent = PassthroughSubject<Event, Never>()
+
+    private init() { }
+
     func notifyEventData(eventData: Event?) {
         guard let eventData = eventData else { return }
-        NotificationCenter.default.post(name: Notification.Name("event"), object: EventDataManager.self, userInfo: ["data": eventData])
+        // NotificationCenter.default.post(name: Notification.Name("event"), object: self, userInfo: ["data": eventData])
+        notifyEvent.send(eventData)
     }
 }
